Fix miter visibility from paint order UI work
The miter limit box should always be visible when the join type is
set to miter. Regardless of how the join type is set. We tightly
couple the miter limit row visibility to the join type updater.

Also remove the Pait Order button type and remove a static pass
thought pattern that was being used to cover a signal.