Move symbol_convert() out of libUEMF
This function is not part of upstream libUEMF, so move it out of our
copy of upstream libUEMF into src/util-string.

Licensing note: although libUEMF is GPLv2-only, the file in question was
added to Inkscape and not to libUEMF, so is assigned GPLv2-or-later in
the absence of clear license information in the file itself (it contains
an ambiguous "Released under GNU GPL").

Additionally,
* Delete an include_directories() in src/CMakeLists.txt.
* Prefer <> includes not containing "3rdparty" for libavoid and libuemf.

Fixes part of https://gitlab.com/inkscape/inkscape/-/issues/973