Set Pango context language to undetermined
This prevents Pango from unconditionally applying a default language to
text.

The Pango language tag is used to apply the OpenType `locl` feature,
which selects language-specific glyphs variants in some fonts. In SVG
the `xml:lang` attribute can be used to specify the language of an
element and its descendants, which Inkscape passes to Pango as the
language tag.

However, when the SVG elements does not have a language specified, the
default behaviour of Pango is to use a default language which is
determined from the current locale of the process. This is undesirable
not only because the default can be different for every users, but also
it prevents the selection of language-specific glyphs by using the font
family name alone (e.g. "Source Han Sans TC" and "Source Han Sans SC").

By setting the context language to `und` (undetermined) we can suppress
this default behaviour so that the default language-specific glyphs of
the font family are used.