Handle `xml:lang`/`lang` attribute updates
The change of language needs to trigger descendant text objects to be
updated, since it is effectively a font change that may change the glyph
selection and even metrics.

The language handling in `SPObject` is refactored such that the
effective language of the object is dynamically inherited from its
parent(s) on demand, instead of cached in the `SPObject` itself.

Added test `test_sp-object-lang` to check the language handling.