Don't set `CMAKE_ENABLE_EXPORTS`
From the CMake docs for `ENABLE_EXPORTS`, this option for executable
targets is intended to make the executables, in case they export
symbols, linkable by other targets. Inkscape does not require this,
therefore the option does not seem to be useful.

This option does not affect debug symbols.

In practice, on Windows CMake adds an `--out-implib` link flag when
linking EXE targets, but the linker may not actually generate an import
library when there are no exported symbols. With the `ENABLE_EXPORTS`
option set, the build system expects the import library to be generated,
but since the file is never generated, the result is that every time an
incremental rebuild is performed, all EXE targets are always being
re-linked even when nothing has actually changed, which is an
undesirable behaviour.