Make global dispatch pool deleted before returning from main
On Windows, non-main threads are terminated before static global
destructors in DLLs get to run. Because of this, if we leave the
destruction of `dispatch_pool` up to static global destructors, a lot of
times it ends up in a deadlock which prevents the Inkscape process from
exiting normally.

To prevent this, we use the `EnableSingleton` system to delete the
`dispatch_pool` shared pointer before returning from `main()`.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/6003