Fix broken filter region calculation
In 'update_filters', setting the blur spinbutton value triggered a 'value_changed' signal, which attempted to update the object, triggering the panel update again.
When a style update occurred (triggering SPStyle::clear), the existing 'filter.href' reference was destroyed and recreated using the SPDocument as the owner, ignoring the specific SPObject that owned the style.
References owned by the document do not register themselves in the target SPFilter's 'hrefList' in the context of item usage. Consequently, after any filter modification, 'hrefList' became empty.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/6007
Fixes: https://gitlab.com/inkscape/inkscape/-/issues/5531